area

+++ src/outputExamples/ageWeight.java
@@ -7,7 +7,7 @@ public class ageWeight {
 	System.out.println("what is your weight");
 	Scanner input = new Scanner(System.in);
 	//your program starts here
-	double allVar = input.nextInt();
+	double allVar = input.nextDouble();
 	System.out.println("you weight " + allVar);
 	System.out.println("what is your height");
 	allVar = input.nextInt();
